# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_pos-retail_go_backend_internal_types_request.CreateGrnRequest:
    properties:
      invoiceNumber:
        type: string
      notes:
        type: string
      poId:
        type: string
      warehouseId:
        type: string
    type: object
...
  github_com_pos-retail_go_backend_internal_types_request.UpdateGrnItemRequest:
    properties:
      id:
        type: string
      orderedQuantity:
        type: integer
      poItemId:
        type: string
      productId:
        type: string
      qualityNotes:
        type: string
      receivedQuantity:
        type: integer
      rejectedQuantity:
        type: integer
      unitPrice:
        type: string
    type: object
  github_com_pos-retail_go_backend_internal_types_request.UpdateGrnRequest:
    properties:
      invoiceNumber:
        type: string
      items:
        items:
          $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_request.UpdateGrnItemRequest'
        type: array
      notes:
        type: string
    type: object
...
  github_com_pos-retail_go_backend_internal_types_response.ApiResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
...
info:
  contact: {}
  description: Go (Fiber) implementation of POS Retail Backend.
  title: POS Retail Backend API
  version: "1.0"
paths:
...
  /api/grn:
    get:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - default: 1
        description: Page
        in: query
        name: page
        type: integer
      - default: 10
        description: Limit
        in: query
        name: limit
        type: integer
      - description: Status
        in: query
        name: status
        type: string
      - description: PO ID
        in: query
        name: poId
        type: string
      - description: Warehouse ID
        in: query
        name: warehouseId
        type: string
      - description: Start date
        in: query
        name: startDate
        type: string
      - description: End date
        in: query
        name: endDate
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
      security:
      - BearerAuth: []
      summary: List GRNs
      tags:
      - GRN
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: GRN payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_request.CreateGrnRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Create GRN
      tags:
      - GRN
  /api/grn/{id}:
    delete:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: GRN ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Cancel GRN
      tags:
      - GRN
    get:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: GRN ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Get GRN
      tags:
      - GRN
    put:
      consumes:
      - application/json
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: GRN ID
        in: path
        name: id
        required: true
        type: string
      - description: Update payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_request.UpdateGrnRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Update GRN
      tags:
      - GRN
  /api/grn/{id}/verify:
    put:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: GRN ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Verify GRN
      tags:
      - GRN
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"